// ClientApp/src/app/record-room/models/ArchivingforPublication.ts
export interface ArchivingforPublication {
  archivingforPublicationId: number;
  departmentNameId:number;
  itemDetailId: number;
  airCraftNameId: number;
  nameofPublicationId: number;
  documentName: string;
  date: Date;
  docUpload: string;
  remarks: string;
  status: number;
  departmentName: string;
  isActive: boolean;
}